fix(auth): use server-side Supabase client for OAuth callback
The auth callback route handler was using createBrowserClient which
cannot set cookies server-side, causing Google login to silently fail
in production. This adds a server-side Supabase client, a session
refresh proxy, fixes a React hydration error on the login page, and
adds Google OAuth config for local development.
