+++ apps/web/components/auth/LoginForm.tsx
@@ -1,0 +1,185 @@
+"use client"
+
+import { useState, useEffect, type FormEvent } from "react"
+import Link from "next/link"
+import { useRouter } from "next/navigation"
+import { useAuth } from "@/lib/data/auth-context"
+import { Button } from "@/components/ui/button"
+import { Input } from "@/components/ui/input"
+
+export default function LoginForm({ initialError }: { initialError: string | null }) {
+  const { login, signInWithApple, signInWithGoogle, profile, isAuthenticated, isLoading } =
+    useAuth()
+  const router = useRouter()
+
+  const [email, setEmail] = useState("")
+  const [password, setPassword] = useState("")
+  const [error, setError] = useState<string | null>(initialError)
+  const [submitting, setSubmitting] = useState(false)
+
+  useEffect(() => {
+    if (!isLoading && isAuthenticated) {
+      router.replace(profile?.onboarding_completed ? "/path" : "/onboarding")
+    }
+  }, [isLoading, isAuthenticated, profile, router])
+
+  if (isLoading || isAuthenticated) return null
+
+  const handleSubmit = async (e: FormEvent) => {
+    e.preventDefault()
+    setError(null)
+
+    if (!email.trim() || !password) {
+      setError("Please fill in all fields.")
+      return
+    }
+
+    setSubmitting(true)
+    try {
+      await login({ email: email.trim(), password })
+    } catch (err) {
+      const message =
+        err instanceof Error ? err.message : "Something went wrong."
+      setError(message)
+      setSubmitting(false)
+    }
+  }
+
+  const handleGoogleSignIn = async () => {
+    setError(null)
+    try {
+      await signInWithGoogle()
+    } catch (err) {
+      const message =
+        err instanceof Error ? err.message : "Something went wrong."
+      setError(message)
+    }
+  }
+
+  const handleAppleSignIn = async () => {
+    setError(null)
+    try {
+      await signInWithApple()
+    } catch (err) {
+      const message =
+        err instanceof Error ? err.message : "Something went wrong."
+      setError(message)
+    }
+  }
+
+  return (
+    <section className="flex min-h-full flex-col items-center justify-center px-6 text-center">
+      <h1 className="text-4xl font-bold tracking-tight">Log in</h1>
+      <p className="mt-3 max-w-sm text-muted-foreground">
+        Welcome back. Pick up where you left off.
+      </p>
+
+      <form
+        onSubmit={handleSubmit}
+        className="mt-8 flex w-full max-w-xs flex-col gap-4"
+        noValidate
+      >
+        <div className="flex flex-col gap-1.5 text-left">
+          <label htmlFor="login-email" className="text-sm font-medium">
+            Email
+          </label>
+          <Input
+            id="login-email"
+            type="email"
+            autoComplete="email"
+            autoCapitalize="none"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            aria-invalid={error ? true : undefined}
+            disabled={submitting}
+          />
+        </div>
+
+        <div className="flex flex-col gap-1.5 text-left">
+          <label htmlFor="login-password" className="text-sm font-medium">
+            Password
+          </label>
+          <Input
+            id="login-password"
+            type="password"
+            autoComplete="current-password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            aria-invalid={error ? true : undefined}
+            disabled={submitting}
+          />
+        </div>
+
+        {error && (
+          <p role="alert" className="text-sm text-destructive">
+            {error}
+          </p>
+        )}
+
+        <Button type="submit" size="lg" disabled={submitting}>
+          {submitting ? "Logging in\u2026" : "Log in"}
+        </Button>
+      </form>
+
+      <div className="mt-4 flex w-full max-w-xs flex-col gap-3">
+        <div className="relative flex items-center justify-center">
+          <span className="absolute inset-x-0 top-1/2 h-px bg-border" />
+          <span className="relative bg-background px-2 text-xs text-muted-foreground">
+            or
+          </span>
+        </div>
+
+        <Button
+          variant="outline"
+          size="lg"
+          onClick={handleAppleSignIn}
+          disabled={submitting}
+          aria-label="Sign in with Apple"
+        >
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="16"
+            height="16"
+            viewBox="0 0 24 24"
+            fill="currentColor"
+            className="mr-2"
+            aria-hidden="true"
+          >
+            <path d="M17.05 20.28c-.98.95-2.05.88-3.08.4-1.09-.5-2.08-.48-3.24 0-1.44.62-2.2.44-3.06-.4C2.79 15.25 3.51 7.59 9.05 7.31c1.35.07 2.29.74 3.08.8 1.18-.24 2.31-.93 3.57-.84 1.51.12 2.65.72 3.4 1.8-3.12 1.87-2.38 5.98.48 7.13-.57 1.5-1.31 2.99-2.54 4.09zM12.03 7.25c-.15-2.23 1.66-4.07 3.74-4.25.29 2.58-2.34 4.5-3.74 4.25z" />
+          </svg>
+          Sign in with Apple
+        </Button>
+
+        <Button
+          variant="outline"
+          size="lg"
+          onClick={handleGoogleSignIn}
+          disabled={submitting}
+          aria-label="Sign in with Google"
+        >
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="16"
+            height="16"
+            viewBox="0 0 24 24"
+            className="mr-2"
+            aria-hidden="true"
+          >
+            <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92a5.06 5.06 0 0 1-2.2 3.32v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.1z" />
+            <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z" />
+            <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" />
+            <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" />
+          </svg>
+          Sign in with Google
+        </Button>
+      </div>
+
+      <Link
+        href="/register"
+        className="mt-6 text-sm text-muted-foreground underline underline-offset-4 hover:text-foreground"
+      >
+        Create account
+      </Link>
+    </section>
+  )
+}
