+++ apps/web/proxy.ts
@@ -1,0 +1,55 @@
+import { createServerClient } from "@supabase/ssr"
+import { NextResponse } from "next/server"
+import type { NextRequest } from "next/server"
+
+export async function proxy(request: NextRequest) {
+  let supabaseResponse = NextResponse.next({ request })
+
+  const url = process.env.NEXT_PUBLIC_SUPABASE_URL
+  const key = process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY
+
+  if (!url || !key) {
+    return supabaseResponse
+  }
+
+  const supabase = createServerClient(url, key, {
+    cookies: {
+      getAll() {
+        return request.cookies.getAll()
+      },
+      setAll(cookiesToSet, headers) {
+        for (const { name, value } of cookiesToSet) {
+          request.cookies.set(name, value)
+        }
+        supabaseResponse = NextResponse.next({ request })
+        for (const { name, value, options } of cookiesToSet) {
+          supabaseResponse.cookies.set(name, value, options)
+        }
+        for (const [headerName, headerValue] of Object.entries(headers)) {
+          supabaseResponse.headers.set(headerName, headerValue)
+        }
+      },
+    },
+  })
+
+  // Refresh the session by verifying the user token.
+  // This triggers a token refresh if needed, writing new cookies via setAll.
+  await supabase.auth.getUser()
+
+  return supabaseResponse
+}
+
+export const config = {
+  matcher: [
+    /*
+     * Match all request paths except:
+     * - _next/static (static files)
+     * - _next/image (image optimization)
+     * - favicon.ico, sitemap.xml, robots.txt (metadata)
+     * - icons/, splash/ (PWA assets)
+     * - sw.js, swe-worker-*.js (service worker)
+     * - manifest.webmanifest
+     */
+    "/((?!_next/static|_next/image|favicon\\.ico|sitemap\\.xml|robots\\.txt|icons/|splash/|sw\\.js|swe-worker-.*\\.js|manifest\\.webmanifest).*)",
+  ],
+}
